+++ packages/ragbits-chat/src/ragbits/chat/providers/model_provider.py
@@ -42,12 +42,11 @@ class RagbitsChatModelProvider:
 
         try:
             from ragbits.chat.auth.types import (
-                CredentialsLoginRequest,
-                JWTToken,
                 LoginRequest,
                 LoginResponse,
-                LogoutRequest,
+                OAuth2AuthorizeResponse,
                 User,
+                UserCredentials,
             )
             from ragbits.chat.interface.forms import UserSettings
             from ragbits.chat.interface.types import (
@@ -73,6 +72,7 @@ class RagbitsChatModelProvider:
                 MessageIdContent,
                 MessageRole,
                 MessageUsage,
+                OAuth2ProviderConfig,
                 Reference,
                 StateUpdate,
                 TextContent,
@@ -123,17 +123,17 @@ class RagbitsChatModelProvider:
                 # API response models
                 "ConfigResponse": ConfigResponse,
                 "FeedbackResponse": FeedbackResponse,
+                "OAuth2AuthorizeResponse": OAuth2AuthorizeResponse,
+                "OAuth2ProviderConfig": OAuth2ProviderConfig,
                 # API request models
                 "ChatRequest": ChatMessageRequest,
                 "FeedbackRequest": FeedbackRequest,
                 # Auth
                 "AuthType": AuthType,
                 "AuthenticationConfig": AuthenticationConfig,
-                "CredentialsLoginRequest": CredentialsLoginRequest,
-                "JWTToken": JWTToken,
+                "UserCredentials": UserCredentials,
                 "LoginRequest": LoginRequest,
                 "LoginResponse": LoginResponse,
-                "LogoutRequest": LogoutRequest,
                 "User": User,
             }
 
@@ -169,7 +169,6 @@ class RagbitsChatModelProvider:
                 "ServerState",
                 "FeedbackItem",
                 "Image",
-                "JWTToken",
                 "User",
                 "MessageUsage",
                 "Task",
@@ -192,16 +191,19 @@ class RagbitsChatModelProvider:
                 "UserSettings",
                 "FeedbackConfig",
                 "AuthenticationConfig",
+                "OAuth2ProviderConfig",
             ],
             "responses": [
                 "FeedbackResponse",
                 "ConfigResponse",
                 "LoginResponse",
+                "OAuth2AuthorizeResponse",
             ],
             "requests": [
                 "ChatRequest",
                 "FeedbackRequest",
-                "CredentialsLoginRequest",
+                "UserCredentials",
+                "OAuth2Credentials",
                 "LoginRequest",
                 "LogoutRequest",
             ],