+++ packages/ragbits-chat/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
+- Change auth backend from jwt to http-only cookie based authentication, add support for OAuth2 authentication (#867)
 
 - Make `SummaryGenerator` optional in `ChatInterface` by providing a default Heuristic implementation.
 - Refactor ragbits-client types to remove excessive use of any (#881)